Feature/azure data analysis

+++ elephantcallscounter/adapters/azure_interface.py
@@ -1,10 +1,14 @@
 import os
+import logging
 import azure.core.exceptions
 from azure.storage.blob import BlobServiceClient
 from azure.storage.blob import BlobClient
 
 from elephantcallscounter.config import env
+from elephantcallscounter.utils.path_utils import get_project_root
 from elephantcallscounter.utils.path_utils import join_paths
+
+logger = logging.getLogger(__name__)
 
 
 class AzureInterface:
@@ -22,19 +26,25 @@ class AzureInterface:
 
         return blob_container_client
 
-    def send_to_azure(self, original_file, dir_path, filename):
-        print('Storing ' + filename + ' in Azure Blob...' + dir_path)
+    def send_to_azure(
+            self, original_file, dir_path, filename, media_file = True, remove_file = False
+    ):
+        logger.info('Storing ' + filename + ' in Azure Blob...' + dir_path)
         try:
             blob_client = self.blob_service_client.get_blob_client(
                 container=self.container_name, blob=join_paths([dir_path, filename])
             )
             with open(original_file, "rb") as data:
-                blob_client.upload_blob(data)
+                if media_file:
+                    blob_client.upload_blob(data, blob_type = 'BlockBlob')
+                else:
+                    blob_client.upload_blob(data)
             # delete local file
-            os.remove(original_file)
-            print('Done uploading file!', dir_path + filename)
+            if remove_file:
+                os.remove(original_file)
+            logger.info('Done uploading file! %s', dir_path + filename)
         except Exception as e:
-            print('Error while uploading ' + filename + ': ' + str(e))
+            logger.info('Error while uploading ' + filename + ': ' + str(e))
 
     def download_from_azure(self, source_file, dest_file):
         try:
@@ -44,9 +54,21 @@ class AzureInterface:
                 blob_name = source_file,
                 credential = env.STORAGE_SAS_KEY
             )
-            with open(os.path.join(dest_file), "wb") as f:
+            with open(join_paths([get_project_root(), dest_file]), "wb+") as f:
                 data = blob.download_blob()
                 data.readinto(f)
         except azure.core.exceptions.ResourceNotFoundError:
-            print('Blob not found', source_file)
+            logger.info('Blob not found %s', source_file)
+
+    def delete_blob(self, blob_name):
+        try:
+            blob = BlobClient(
+                account_url = env.AZURE_STORAGE_ACCOUNT,
+                container_name = self.container_name,
+                blob_name = blob_name,
+                credential = env.STORAGE_SAS_KEY
+            )
+            blob.delete_blob()
+        except azure.core.exceptions.ResourceNotFoundError:
+            logger.info('Blob not found %s', blob_name)
